requirements.txt
streamlit
openai
langchain
langchain-community
langchain-core
python-dotenv
